OnboardingTermsScreen 리팩터링 및 위치 이동
- `OnboardingTermsScreen.kt` 파일을 `terms` 패키지에서 상위 패키지로 이동
- `OnboardingStepScaffold`를 적용하여 공통 레이아웃 구조 통합 및 코드 간소화
- `TermsType` 열거형을 도입하여 약관 상세 보기 클릭 이벤트 처리 방식 개선
- `TermsState`에서 불필요한 `isNextEnabled` 계산 로직 제거
- `TermsRow` 컴포저블의 레이아웃 및 스타일(Typography) 수정으로 UI 일관성 강화
- `HorizontalDivider`를 사용하여 약관 항목 간 구분선 구현 방식 변경
